tabular_data/ember/requirements.txt
pandas==1.0.3
tqdm==4.42.1
lightgbm==2.3.1
numpy==1.18.2
requests==2.19.1
lief==0.10.1
alectio_sdk==0.0.1
keras==2.4.3
PyYAML==5.3.1
scikit_learn==0.23.1
tensorflow==2.2.0
